requirements.txt
aiofiles
websockets
azure-storage-blob
python-dotenv
aiohttp
requests
websocket
python-dotenv
websockets
requests
websocket
websocket-client